Ejercicio propuesto numero 25.

+++ src/EjerciciosPropuestos/VentiunoAlTreinta/Propuesto25.java
@@ -20,7 +20,7 @@ public class Propuesto25 {
         */
 
         //Variables
-        double sueldo, descuento = 0;
+        int sueldo, descuento = 0;
         char trabajador, sexo;
